requirements.txt
fastapi
uvicorn
openai
httpx
beautifulsoup4
python-dotenv
json5
tiktoken
pydantic
requests
pyyaml
